APLICACAO_WEB/requirements.txt
asgiref==3.5.2
astroid==2.11.7
bcrypt==4.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.4
debugpy==1.6.4
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.6
Django==3.2.16
django-extra-views==0.14.0
filelock==3.8.0
gcloud==0.18.3
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
isort==5.10.1
jwcrypto==1.3.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Naked==0.1.31
oauth2client==4.1.3
paramiko==3.1.0
platformdirs==2.5.2
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.15.0
pylint==2.14.4
PyNaCl==1.5.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==3.3.3
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.16.5
wrapt==1.14.1
